pyproject.toml
[tool.poetry]
name = "repl_python3_Spirit"
version = "0.1.0"
description = ""
authors = ["YTJurassic <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^2.0.1"
python-dotenv = "^0.17.1"
discord = "^1.0.1"
replit = "^3.1.0"
youtube-dl = "^2021.5.16"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
